+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.7.1] - 2025-12-08
+
 ### Added
 - Implement mandatory config check on command execution [SCI-23]
   - Added global pre-execution check that aborts non-whitelisted commands when config file is missing